break schema out into a subpackage
